refactor directories
- added different folders in view for every use case
- fixed redirection to statistics when user finishes playing the quiz by adding a default user
