+++ lib/src/view/play_quiz/play_quiz_usecase.dart
@@ -1,0 +1,3 @@
+import 'package:flutter/material.dart';
+
+class PlayQuizUseCase {}
