+++ lib/src/view/play_quiz/quiz_view.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:queasy/constants/app_themes.dart';
-import 'package:queasy/src/controller/quiz_view_controller.dart';
+import 'package:queasy/src/view/play_quiz/quiz_view_controller.dart';
 import 'package:queasy/src/view/statistics_view.dart';
 import 'package:queasy/src/view/widgets/custom_bottom_nav_bar.dart';
