+++ lib/src/view/see_profile/profile_view.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:queasy/constants/app_themes.dart';
-import 'package:queasy/src/controller/profile-view-controller.dart';
+import 'package:queasy/src/view/see_profile/profile-view-controller.dart';
 import 'package:queasy/src/view/login_view.dart';
-import 'widgets/custom_bottom_nav_bar.dart';
+import '../widgets/custom_bottom_nav_bar.dart';
 
 class UserProfile extends StatefulWidget {
   final ProfileViewController controller = ProfileViewController();
